Some blueprint refactoring and testing
